Added Parameter to Remove Extraneous Files on Upload
This commit adds two parameters (--delete, --delete-excluded) to the upload
command. When specified, files that have been removed locally will be removed
(more precisely will be moved to the trash) on the server as well.

This functionality is usefull to keep local and remote conent in sync. It mimics
the behaviour of the respective rsync commands.